+++ packages/paste-core/utilities/grid/stories/index.stories.tsx
@@ -3,6 +3,10 @@ import {storiesOf} from '@storybook/react';
 import {withKnobs, select} from '@storybook/addon-knobs';
 import {DefaultTheme, ThemeShape} from '@twilio-paste/theme-tokens';
 import {Box} from '@twilio-paste/box';
+import {Card} from '@twilio-paste/card';
+import {Heading} from '@twilio-paste/heading';
+import {Paragraph} from '@twilio-paste/paragraph';
+import {Text} from '@twilio-paste/text';
 import {Grid, Column} from '../src';
 
 const spaceOptions = Object.keys(DefaultTheme.space);
@@ -14,62 +18,62 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter={gutterValue}>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             2
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             3
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             4
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             5
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             6
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             7
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             8
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             9
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             10
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             11
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             12
           </Box>
         </Column>
@@ -80,7 +84,7 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
@@ -91,12 +95,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             2
           </Box>
         </Column>
@@ -107,17 +111,17 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             2
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             3
           </Box>
         </Column>
@@ -128,22 +132,22 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             2
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             3
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             4
           </Box>
         </Column>
@@ -154,32 +158,32 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             2
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             3
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             4
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             5
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             6
           </Box>
         </Column>
@@ -190,17 +194,17 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={5}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             5
           </Box>
         </Column>
         <Column span={5}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             5
           </Box>
         </Column>
         <Column span={2}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             2
           </Box>
         </Column>
@@ -211,12 +215,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={6}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             6
           </Box>
         </Column>
         <Column span={6}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             6
           </Box>
         </Column>
@@ -227,12 +231,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={8}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             8
           </Box>
         </Column>
         <Column span={4}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             4
           </Box>
         </Column>
@@ -243,12 +247,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={9}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             9
           </Box>
         </Column>
         <Column span={3}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             3
           </Box>
         </Column>
@@ -259,12 +263,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={10}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             10
           </Box>
         </Column>
         <Column span={2}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             2
           </Box>
         </Column>
@@ -275,12 +279,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={11}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             11
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
@@ -291,57 +295,57 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={2}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             2
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
@@ -352,27 +356,27 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={8}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             8
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
         <Column span={1}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             1
           </Box>
         </Column>
@@ -383,7 +387,7 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid>
         <Column span={[12, 6, 8]} offset={[0, 3, 2]}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             8
           </Box>
         </Column>
@@ -394,12 +398,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={[12, 6, 6]} offset={[0, 2, 2]}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             6
           </Box>
         </Column>
         <Column span={4}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             4
           </Box>
         </Column>
@@ -410,12 +414,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid>
         <Column span={4}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             4
           </Box>
         </Column>
         <Column span={4} offset={4}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             4
           </Box>
         </Column>
@@ -426,12 +430,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid>
         <Column span={2} offset={2}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             2
           </Box>
         </Column>
         <Column span={2} offset={6}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             2
           </Box>
         </Column>
@@ -442,12 +446,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20">
         <Column span={3}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             3
           </Box>
         </Column>
         <Column span={9}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             9
           </Box>
         </Column>
@@ -458,17 +462,17 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space60">
         <Column span={3}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             3
           </Box>
         </Column>
         <Column span={9}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             9
           </Box>
         </Column>
         <Column span={3}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             3
           </Box>
         </Column>
@@ -479,12 +483,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter={['space20', 'space60', 'space90']}>
         <Column span={[2, 6, 8]}>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             Responsive Column Size
           </Box>
         </Column>
         <Column span={[10, 6, 4]}>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             Responsive Column Size
           </Box>
         </Column>
@@ -495,12 +499,12 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space20" vertical={[true, false, false]}>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             Responsive Column Size
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
             Responsive Column Size
           </Box>
         </Column>
@@ -511,20 +515,20 @@ storiesOf('Utilities|Grid', module)
     return (
       <Grid gutter="space40">
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             1
           </Box>
         </Column>
         <Column>
-          <Box height="size40" width="100%">
+          <Box height="size40">
             <Grid gutter="space20">
               <Column>
-                <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+                <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
                   Nested Column 1
                 </Box>
               </Column>
               <Column>
-                <Box backgroundColor="colorBackgroundPrimaryLight" height="size40" width="100%">
+                <Box backgroundColor="colorBackgroundPrimaryLight" height="size40">
                   Nested Column 2
                 </Box>
               </Column>
@@ -532,9 +536,57 @@ storiesOf('Utilities|Grid', module)
           </Box>
         </Column>
         <Column>
-          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40" width="100%">
+          <Box backgroundColor="colorBackgroundPrimaryLighter" height="size40">
             3
           </Box>
+        </Column>
+      </Grid>
+    );
+  })
+  .add('Grid - 3 Column Card Layout', () => {
+    return (
+      <Grid gutter="space70" vertical={[true, false, false]}>
+        <Column>
+          <Card padding="space70">
+            <Heading as="h2">Card Heading</Heading>
+            <Text as="p">Body</Text>
+          </Card>
+        </Column>
+        <Column>
+          <Card padding="space70">
+            <Heading as="h2">Card Heading</Heading>
+            <Text as="p">Body</Text>
+          </Card>
+        </Column>
+        <Column>
+          <Card padding="space70">
+            <Heading as="h2">Card Heading</Heading>
+            <Text as="p">Body</Text>
+          </Card>
+        </Column>
+      </Grid>
+    );
+  })
+  .add('Grid - 2 Column Content with Card', () => {
+    return (
+      <Grid gutter="space70" vertical={[true, false, false]}>
+        <Column span={8}>
+          <Heading as="h2">Content Heading</Heading>
+          <Paragraph>
+            Kale chips distillery authentic, portland etsy cloud bread vinyl gentrify drinking vinegar viral meh hot
+            chicken bitters fashion axe palo santo. Chillwave fixie sustainable <i>helvetica</i> etsy.
+          </Paragraph>
+          <Paragraph>
+            Prism whatever ethical, gochujang <strong>edison bulb</strong> put a bird on it kitsch. Pop-up 90&apos;s la
+            croix tumeric, palo santo chia try-hard direct trade tote bag roof party scenester kitsch stumptown
+            intelligentsia. Literally heirloom blue bottle etsy.
+          </Paragraph>
+        </Column>
+        <Column span={4}>
+          <Card padding="space70">
+            <Heading as="h2">Card Heading</Heading>
+            <Text as="p">Body</Text>
+          </Card>
         </Column>
       </Grid>
     );