+++ packages/paste-core/core-bundle/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.11](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@2.1.10...@twilio-paste/core@2.1.11) (2020-02-11)
+
+**Note:** Version bump only for package @twilio-paste/core
+
+
+
+
+
+## [2.1.10](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@2.1.9...@twilio-paste/core@2.1.10) (2020-02-11)
+
+**Note:** Version bump only for package @twilio-paste/core
+
+
+
+
+
 ## [2.1.9](https://github.com/twilio-labs/paste/compare/@twilio-paste/core@2.1.8...@twilio-paste/core@2.1.9) (2020-02-11)
 
 **Note:** Version bump only for package @twilio-paste/core